requirements.txt
Django==5.2.7
djangorestframework
djangorestframework-simplejwt
celery[redis]
redis
Pillow
requests
numpy
torch